For libfabric-sys, added libfabric C library as a submodule

+++ build.rs
@@ -6,14 +6,14 @@ fn main(){
     
     let out_path = std::path::PathBuf::from(std::env::var("OUT_DIR").unwrap());
     let src_path = std::fs::canonicalize(std::path::PathBuf::from("libfabric")).unwrap();
-    let build_path = src_path.join("build");
+    let build_path = out_path.join("build");
     let custom_path = match std::env::var("LIBFABRIC_ROOT") {
         Ok(val) => val,
         Err(_) => "".to_string(),
     };
     
     let libfabrics_path;
-    if custom_path == ""{
+    if custom_path.is_empty(){
         
         libfabrics_path = build_path.to_str().unwrap().to_owned();
         // println!("cargo:rerun-if-changed={}", libfabrics_path.clone()+"/include/fabric_sys.h");
@@ -31,7 +31,7 @@ fn main(){
         if !build_path.exists(){
     
             std::process::Command::new("sh").current_dir(src_path.to_str().unwrap()).arg("autogen.sh").status().unwrap();
-            std::process::Command::new("sh").current_dir(src_path.to_str().unwrap()).arg("configure").arg("--prefix=".to_owned()+build_path.to_str().unwrap()).arg("--enable-static").status().unwrap();
+            std::process::Command::new("sh").current_dir(src_path.to_str().unwrap()).arg("configure").arg("CFLAGS=-fPIC").arg("--prefix=".to_owned()+build_path.to_str().unwrap()).arg("--enable-static").status().unwrap();
         }
         
         
@@ -46,56 +46,61 @@ fn main(){
     
     // libfabric has multiple header file that we are required to generate bindings for so we create a
     // new header named fabric_sys.h and #include all headers there
-    let header_file = std::fs::File::create(out_path.to_str().unwrap().to_owned()+"/fabric_sys.h").unwrap();
-    let mut writer = std::io::BufWriter::new(header_file);
-    
-    // Another problem is that there are several static inline functions in these header files which do
-    // not create symbols in the libfabric.so and thus cannot be linked from rust.
-    // For this reason, we create wrapper functions for each inline one and store them in a file called inlined.c
-    // which resides in the /build/ directory 
-    let inlined_file = std::fs::File::create(out_path.to_str().unwrap().to_owned()+"/inlined.c").unwrap();
-    let mut writer_inlined = std::io::BufWriter::new(inlined_file);
-    let _ = writer_inlined.write_all(b"#include<fabric_sys.h>\n");
-    
-    // We keep the prototypes of all the wrappers fo rinline functions here and will append them to fabric_sys.h
-    // so that they are visible to bindgen
-    let mut inlined_funcs: Vec<String> = Vec::new();
+    let header_path = std::path::PathBuf::from(out_path.to_str().unwrap().to_owned()+"/fabric_sys.h");
+    let lib_path = out_path.join("libinlined.so");
+    if !header_path.exists() || !lib_path.exists(){
 
-    let headers = std::fs::read_dir(libfabrics_path.clone()+"/include/rdma/").unwrap();
-    
-    // For each file in libfabric/include/
-    for header in headers {
-        if header.as_ref().unwrap().file_type().unwrap().is_file() {
-            // Create the wrappers (prototype, definition) for all inlined functions
-            let mut funcs = crate::create_inlined_wrappers::read_file(header.as_ref().unwrap().path().to_str().unwrap());
-            
-            // Store the prototype for later
-            inlined_funcs.append(&mut funcs.0);
-            
-            // Write the definition to inlined.c
-            for f in funcs.1 {
-                let _ = writer_inlined.write_all((f+"\n").as_bytes());
+        let header_file = std::fs::File::create(out_path.to_str().unwrap().to_owned()+"/fabric_sys.h").unwrap();
+        let mut writer = std::io::BufWriter::new(header_file);
+        
+        // Another problem is that there are several static inline functions in these header files which do
+        // not create symbols in the libfabric.so and thus cannot be linked from rust.
+        // For this reason, we create wrapper functions for each inline one and store them in a file called inlined.c
+        // which resides in the /build/ directory 
+        let inlined_file = std::fs::File::create(out_path.to_str().unwrap().to_owned()+"/inlined.c").unwrap();
+        let mut writer_inlined = std::io::BufWriter::new(inlined_file);
+        let _ = writer_inlined.write_all(b"#include<fabric_sys.h>\n");
+        
+        // We keep the prototypes of all the wrappers fo rinline functions here and will append them to fabric_sys.h
+        // so that they are visible to bindgen
+        let mut inlined_funcs: Vec<String> = Vec::new();
+        
+        let headers = std::fs::read_dir(libfabrics_path.clone()+"/include/rdma/").unwrap();
+        
+        // For each file in libfabric/include/
+        for header in headers {
+            if header.as_ref().unwrap().file_type().unwrap().is_file() {
+                // Create the wrappers (prototype, definition) for all inlined functions
+                let mut funcs = crate::create_inlined_wrappers::read_file(header.as_ref().unwrap().path().to_str().unwrap());
+                
+                // Store the prototype for later
+                inlined_funcs.append(&mut funcs.0);
+                
+                // Write the definition to inlined.c
+                for f in funcs.1 {
+                    let _ = writer_inlined.write_all((f+"\n").as_bytes());
+                }
+                
+                // #include the header to fabric_sys.h
+                let _ = writer.write_all(("#include<rdma/".to_owned()+&header.as_ref().unwrap().file_name().into_string().unwrap()+">\n").as_bytes());
             }
-
-            // #include the header to fabric_sys.h
-            let _ = writer.write_all(("#include<rdma/".to_owned()+&header.as_ref().unwrap().file_name().into_string().unwrap()+">\n").as_bytes());
         }
+        
+        // Append the prototypes of the wrappers for the inline functions to fabric_sys.h
+        for f in inlined_funcs {
+            let _ = writer.write_all((f+"\n").as_bytes());
+        }
+        
+        // Make sure the files are written completely
+        let _ = writer.flush();
+        let _ = writer_inlined.flush();
+        
+        // Create a new lib, libinlined.so that we expose the wrappers for rust to use
+        println!("{}", out_path.to_str().unwrap());
+        println!("{}", libfabrics_path);
+        std::process::Command::new("gcc").current_dir(out_path.to_str().unwrap()).arg("-Wno-everything").arg("-O3").arg("-I".to_owned()+&libfabrics_path+"/include/").arg("-I".to_owned()+out_path.to_str().unwrap()).arg("--shared").arg("-fPIC").arg("-o").arg("libinlined.so").arg(out_path.to_str().unwrap().to_owned()+"/inlined.c").status().unwrap();
     }
-
-    // Append the prototypes of the wrappers for the inline functions to fabric_sys.h
-    for f in inlined_funcs {
-        let _ = writer.write_all((f+"\n").as_bytes());
-    }
-
-    // Make sure the files are written completely
-    let _ = writer.flush();
-    let _ = writer_inlined.flush();
-
-    // Create a new lib, libinlined.so that we expose the wrappers for rust to use
-    println!("{}", out_path.to_str().unwrap());
-    println!("{}", libfabrics_path);
-    std::process::Command::new("gcc").current_dir(out_path.to_str().unwrap().to_owned()).arg("-Wno-everything").arg("-O3").arg("-I".to_owned()+&libfabrics_path+"/../include/").arg("-I".to_owned()+out_path.to_str().unwrap()).arg("--shared").arg("-fPIC").arg("-o").arg("libinlined.so").arg(out_path.to_str().unwrap().to_owned()+"/inlined.c").status().unwrap();
-    // std::process::Command::new("gcc").current_dir(out_path.to_str().unwrap().to_owned()).arg("-Wno-everything").arg("-O3").arg("-I".to_owned()+&libfabrics_path+"/../include/").arg("-I".to_owned()+out_path.to_str().unwrap()).arg("-c").arg("-o").arg("inlined.o").arg(out_path.to_str().unwrap().to_owned()+"/inlined.c").status().unwrap();
+    // std::process::Command::new("gcc").current_dir(out_path.to_str().unwrap().to_owned()).arg("-Wno-everything").arg("-fPIC").arg("-O3").arg("-I".to_owned()+&libfabrics_path+"/include/").arg("-I".to_owned()+out_path.to_str().unwrap()).arg("-c").arg("-o").arg("inlined.o").arg(out_path.to_str().unwrap().to_owned()+"/inlined.c").status().unwrap();
     // std::process::Command::new("ar").current_dir(out_path.to_str().unwrap().to_owned()).arg("-rc").arg("libinlined.a").arg("inlined.o").status().unwrap();
     // Generate the rust bindings
     let bindings = bindgen::Builder::default()
